requirements.txt
PyJWT
requests
Flask
Flask-RESTful
Flask-JWT==0.3.2
